fix for fork to be compatible with fork

+++ Source/GSXibLoader.m
@@ -328,8 +328,8 @@
             { 	 
               const char *nam = [label cString]; 	 
               const char *type; 	 
-              unsigned int size; 	 
-              int offset; 	 
+              size_t	 size; 	 
+              ptrdiff_t offset; 	 
               
               /* 	 
                * Use the GNUstep additional function to set the instance
